% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LapTimeToSeconds.R
\name{lap_time_to_seconds}
\alias{lap_time_to_seconds}
\title{Convert lap time strings to seconds}
\usage{
lap_time_to_seconds(lap_time_vector)
}
\arguments{
\item{lap_time_vector}{A character vector of lap time strings}
}
\value{
A numeric vector of lap times converted to total seconds
}
\description{
This function takes a vector of lap time strings and converts each element
to the total number of seconds. It handles special cases like "NaT" (Not a Time)
and lap times with a "0 days" prefix.
}
\details{
The function supports lap time formats like "HH:MM:SS" or "0 days HH:MM:SS".
It returns a numeric vector of the same length as the input, where each element
represents the total seconds for the corresponding lap time.
}
\examples{
lap_times <- c("01:23:45", "00:59:59", "NaT", "0 days 02:30:00")
lap_time_to_seconds(lap_times)
}
